% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCSBpckg2.R
\name{detect_delim}
\alias{detect_delim}
\title{scCSBpckg2}
\usage{
detect_delim(file_path)
}
\value{
A matrix containing the data from the CSV, loom, h5ad and cellRanger scRNA files.
}
\description{
scCSBpckg2
}
\examples{
#csv1 <- sc_data_import("./tests/Test_Files/csv_test.csv")
#print(csv1)
#loom1 <- sc_data_import("./tests/Test_Files/loom_test.loom")
#print(loom1)
#h5ad1 <- sc_data_import("./tests/Test_Files/h5ad_test.h5ad")
#print(h5ad1)
#txt <- sc_data_import("./tests/Test_Files/txt_test.txt")
#print(txt)
#tenx <- sc_data_import("./tests/Test_Files/cellRanger_test/")
#print(tenx)

}
